pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">= 3.12"
version = "0.1"
dependencies = [
    'tinkoff-investments',
    'aiofiles',
    'python-dotenv',
    'tzdata',
]
description = 'Wrapper package for tinkoff-investments'